+++ src/main/java/org/synyx/urlaubsverwaltung/availability/api/AvailabilityApiController.java
@@ -4,9 +4,10 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -19,9 +20,9 @@ import java.time.LocalDate;
 import java.util.Optional;
 
 
-@Api("Availabilities: Get all availabilities for a certain period")
+@Api("Availabilities: Get all availabilities for a certain person and period")
 @RestController("restApiAvailabilityController")
-@RequestMapping("/api/")
+@RequestMapping("/api/persons/{personId}")
 public class AvailabilityApiController {
 
     public static final String AVAILABILITIES = "availabilities";
@@ -43,17 +44,16 @@ public class AvailabilityApiController {
     @GetMapping(AVAILABILITIES)
     @PreAuthorize(SecurityRules.IS_OFFICE)
     public AvailabilityListDto personsAvailabilities(
-        @ApiParam(value = "start of interval to get availabilities from (inclusive)", defaultValue = RestApiDateFormat.EXAMPLE_FIRST_DAY_OF_YEAR)
+        @ApiParam(value = "id of the person")
+        @PathVariable(value = "personId")
+            Integer personId, @ApiParam(value = "start of interval to get availabilities from (inclusive)", defaultValue = RestApiDateFormat.EXAMPLE_FIRST_DAY_OF_YEAR)
         @RequestParam("from")
         @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
             LocalDate startDate,
         @ApiParam(value = "end of interval to get availabilities from (inclusive)", defaultValue = RestApiDateFormat.EXAMPLE_LAST_DAY_OF_MONTH)
         @RequestParam("to")
         @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
-            LocalDate endDate,
-        @ApiParam(value = "id of the person")
-        @RequestParam(value = "person")
-            Integer personId) {
+            LocalDate endDate) {
 
         final Optional<Person> optionalPerson = personService.getPersonByID(personId);